+++ Client.h
@@ -14,7 +14,7 @@ friend std::ostream& operator<<(std::ostream& os, const Client& Client);
 private:
     int id; /**<ID relacionado a la tarjeta RFID unica de cada cliente*/
     Vehicle vehicle; /**<Objeto que almacena los datos del vehiculo*/
-    float balance; /**<Lleva el saldo del cliente, en pesos argentinos*/
+    double balance; /**<Lleva el saldo del cliente, en pesos argentinos*/
 
 public:
     Client( int id = 0, const std::string& name ="", int age = 0, const std::string& address ="", const std::string& email = "",
@@ -22,12 +22,12 @@ public:
             const std::string& color = "", const std::string& brand = "", const std::string& model = "");
 
     int getId() const;
-    int getBalance() const;
+    double getBalance() const;
     Vehicle getVehicle() const;
 
     void setId(int id);
     void setVehicle(const Vehicle& vehicle);
-    void setBalance(float balance);
+    void setBalance(double balance);
 };
 
 #endif
